DIG-92 Moved api core functionality to ExportService
